# perf: upgrade compress package test and benchmark. 