--- a/go.mod
+++ b/go.mod
@@ -8,9 +8,8 @@ require (
 	github.com/dustin/go-humanize v1.0.0
 	github.com/gogo/protobuf v1.3.2
 	github.com/golang/protobuf v1.5.2
-	github.com/golang/snappy v0.0.3
 	github.com/google/flatbuffers v1.12.1
-	github.com/klauspost/compress v1.12.3
+	github.com/klauspost/compress v1.15.15
 	github.com/pkg/errors v0.9.1
 	github.com/spf13/cobra v1.7.0
 	github.com/stretchr/testify v1.4.0